Rework cost as BigDecimal

+++ unit-10/src/main/java/io/github/xisyn/spring/data/service/PurchasedBookService.java
@@ -4,13 +4,15 @@ import io.github.xisyn.spring.data.entity.Book;
 import io.github.xisyn.spring.data.entity.Customer;
 import io.github.xisyn.spring.data.entity.PurchasedBook;
 
+import java.math.BigDecimal;
+
 public interface PurchasedBookService {
 
     void save(PurchasedBook purchasedBook);
 
-    void purchaseBook(Book book, Customer customer, double cost);
+    void purchaseBook(Book book, Customer customer, BigDecimal cost);
 
-    double bookAmountPurchases(Book book);
+    BigDecimal bookAmountPurchases(Book book);
 
-    double customerAmountPurchases(Customer customer);
+    BigDecimal customerAmountPurchases(Customer customer);
 }
